image-service/requirements.txt
fastapi==0.103.0
uvicorn==0.22.0
python-multipart==0.0.6
numpy==1.26.4
rasterio==1.3.6
matplotlib==3.8.1
pyjwt[crypto]==2.7.0